+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/controller/GrowthProcessController.java
@@ -25,7 +25,7 @@ public class GrowthProcessController {
 
     @GetMapping
     public ResponseEntity<ApiResponse<List<GrowthProcessDto>>> getAll(
-            @RequestParam(required = false) Long cropId) {
+            @RequestParam(required = false) Integer cropId) {
         try {
             List<GrowthProcessDto> data = (cropId != null)
                     ? growthProcessService.findByCropId(cropId)
@@ -38,7 +38,7 @@ public class GrowthProcessController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<ApiResponse<GrowthProcessDto>> getById(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<GrowthProcessDto>> getById(@PathVariable Integer id) {
         return growthProcessService.findById(id)
                 .map(data -> ResponseEntity.ok(ApiResponse.success(data)))
                 .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND)
@@ -57,7 +57,7 @@ public class GrowthProcessController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ApiResponse<GrowthProcessDto>> update(@PathVariable Long id,
+    public ResponseEntity<ApiResponse<GrowthProcessDto>> update(@PathVariable Integer id,
             @RequestBody GrowthProcessDto dto) {
         try {
             return growthProcessService.update(id, dto)
@@ -71,7 +71,7 @@ public class GrowthProcessController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse<Void>> delete(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<Void>> delete(@PathVariable Integer id) {
         try {
             if (growthProcessService.deleteById(id)) {
                 return ResponseEntity.ok(ApiResponse.success(null, "GrowthProcess deleted successfully"));
@@ -85,7 +85,7 @@ public class GrowthProcessController {
     }
 
     @DeleteMapping("/bulk-delete")
-    public ResponseEntity<ApiResponse<Void>> deleteBulk(@RequestBody List<Long> ids) {
+    public ResponseEntity<ApiResponse<Void>> deleteBulk(@RequestBody List<Integer> ids) {
         try {
             int deletedCount = growthProcessService.deleteByIds(ids);
 
@@ -98,7 +98,7 @@ public class GrowthProcessController {
     }
 
     @GetMapping("/detail/{id}")
-    public ResponseEntity<ApiResponse<GrowthProcessDetailDto>> getDetailById(@PathVariable Long id) {
+    public ResponseEntity<ApiResponse<GrowthProcessDetailDto>> getDetailById(@PathVariable Integer id) {
         return growthProcessService.findDetailById(id)
                 .map(data -> ResponseEntity.ok(ApiResponse.success(data)))
                 .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND)
