+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/dto/InventoryBatchDto.java
@@ -9,8 +9,8 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class InventoryBatchDto {
-    private Long batchInvId;
-    private Long itemId;
+    private Integer batchInvId;
+    private Integer itemId;
     private String itemName;
     private String supplier;
     private Double quantity;
