+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/dto/PlantingBatchDto.java
@@ -8,12 +8,12 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class PlantingBatchDto {
-    private Long pBatchId;
+    private Integer pBatchId;
     private String batchName;
-    private Long cropId;
+    private Integer cropId;
     private String cropName;
     private String variety;
-    private Long processId;
+    private Integer processId;
     private String processName;
     private Double areaM2;
     private Integer cropCount;
