+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/entity/Task.java
@@ -16,7 +16,7 @@ public class Task {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "task_id")
-    private Long taskId;
+    private Integer taskId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "p_batch_id")
