+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/repository/InventoryBatchRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface InventoryBatchRepository extends JpaRepository<InventoryBatch, Long> {
-    List<InventoryBatch> findByItemItemId(Long itemId);
+public interface InventoryBatchRepository extends JpaRepository<InventoryBatch, Integer> {
+    List<InventoryBatch> findByItemItemId(Integer itemId);
 }
