+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/repository/TaskRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TaskRepository extends JpaRepository<Task, Long> {
+public interface TaskRepository extends JpaRepository<Task, Integer> {
 
     // Sử dụng @Query để viết rõ câu lệnh JPQL
     @Query("SELECT t FROM Task t " +
@@ -17,16 +17,16 @@ public interface TaskRepository extends JpaRepository<Task, Long> {
             "LEFT JOIN FETCH tt.stage s " +
             "WHERE t.plantingBatch.pBatchId = :batchId " +
             "ORDER BY s.stageId, t.plannedDate ASC")
-    List<Task> findTasksWithStageByBatchId(@Param("batchId") Long batchId);
+    List<Task> findTasksWithStageByBatchId(@Param("batchId") Integer batchId);
 
     @Query("SELECT t FROM Task t WHERE t.plantingBatch.pBatchId = :batchId")
-    List<Task> findByBatchId(@Param("batchId") Long batchId);
+    List<Task> findByBatchId(@Param("batchId") Integer batchId);
 
-    List<Task> findByAssignedToUserId(Long userId);
+    List<Task> findByAssignedToUserId(Integer userId);
 
     List<Task> findByStatus(String status);
 
     // Nếu hàm này cũng lỗi, bạn có thể sửa tương tự:
     @Query("SELECT t FROM Task t WHERE t.plantingBatch.pBatchId = :batchId ORDER BY t.plannedDate ASC")
-    List<Task> findTasksByBatchIdOrderByPlannedDateAsc(@Param("batchId") Long batchId);
+    List<Task> findTasksByBatchIdOrderByPlannedDateAsc(@Param("batchId") Integer batchId);
 }
